Connect frontend with Firebase Auth
This connects the frontend (sign in and asign up) to Firebase auth using the email and password method.
The app now keeps track of user context, i.e if the user is signed in or not. If the user is signed in they are redirected
to the home page. If not, they will be at the login page.